desktop_app/requirements.txt
pyserial
tkinter
matplotlib
numpy
